requirements.txt
anthropic>=0.52.0
python-pptx>=1.0.0
striprtf>=0.0.26
google-genai>=0.8.0
Pillow>=10.0.0